Fix chr() bigint cast error in character-name backfill migration
row_number() returns bigint but Postgres's chr() only accepts int,
so the not-yet-applied migration failed with "function chr(bigint)
does not exist" when the user ran it. Cast to int explicitly.

+++ supabase/migrations/20260727090000_names_stacks_deletion.sql
@@ -24,7 +24,7 @@ with numbered as (
   where name is null
 )
 update public.characters c
-set name = 'Legacy' || chr(96 + numbered.rn)
+set name = 'Legacy' || chr((96 + numbered.rn)::int)
 from numbered
 where c.id = numbered.id;
